requirements.txt
pandas
numpy
matplotlib
networkx
scipy
python-louvain
pyvis
torch
torch-geometric
scikit-learn
xgboost
seaborn
umap-learn
tqdm